Convert to websocket_client library.  Basic message handling
Why:

We need a better way to manage the lifecycle of the bot that is more in
line with OTP.  Also we want to allow users to easily start adding new
handlers.

This change addresses the need by:

I've swapped out the old socket library for the erlang client.  In the
future it would be great to use a pure elixir version but that version
does not currently exist.  It might be something that we could write
ourselves.  I've also added a very basic message handler so that Butler
can respond to messages.  The handle message api isn't as friendly as it
could be.  However, it allows for more flexibility when writing specific
message handlers.  Once we have a feel for the handlers I'd like to try
to create a more intuitive api (including regex parsing and not having
to know as much about the underlying methodology).

We're capturing the initial slack org state when teh bot starts up.
However, we don't update that state based on any of the messages.  Now
that we have this skeleton in place we can start adding specific
listeners to handle slack specific messages and use those messages to
keep our internal bot state up to date.