Refactor plugin architecture
Why:

Plugins should run for indeterminant amounts of time and also be
non-blocking.

This change addresses the need by:

Plugins now get their own task to run in. Plugins have also been
refactored to take a connection object (really a `Butler.Message`). This
struct contains everything that the plugin needs in order to respond to
the proper user and channel. The plugin then responds with a
`Butler.Response` built from the text response and from the data passed
in the initial connection struct.